fix build error on new android studio

+++ Vins-Droidcv310/app/src/main/cpp/src/vins_estimator/src/parameters.cpp
@@ -52,7 +52,7 @@ int STEREO_TRACK;
 int EQUALIZE;
 int ROW;
 int COL;
-int FOCAL_LENGTH;
+double FOCAL_LENGTH;
 int FISHEYE;
 bool PUB_THIS_FRAME;
